chore: swap TxDialog with TransactionDialog component
