+++ src/app/pages/LiquidityMining/components/RemoveLiquidityDialog/RemoveLiquidityDialogV1.tsx
@@ -20,7 +20,6 @@ import { DummyInput } from 'app/components/Form/Input';
 import { AssetRenderer } from '../../../../components/AssetRenderer';
 import { TxFeeCalculator } from '../../../MarginTradePage/components/TxFeeCalculator';
 import { DialogButton } from 'app/components/Form/DialogButton';
-import { TxDialog } from '../../../../components/Dialogs/TxDialog';
 import { LoadableValue } from '../../../../components/LoadableValue';
 import { weiToNumberFormat } from '../../../../../utils/display-text/format';
 import { useLiquidityMining_getUserInfo } from '../../hooks/useLiquidityMining_getUserInfo';
@@ -34,6 +33,7 @@ import { discordInvite } from 'utils/classifiers';
 import { AmmLiquidityPool } from 'utils/models/amm-liquidity-pool';
 
 import { useCacheCallToWithValue } from 'app/hooks/chain/useCacheCallToWithValue';
+import { TransactionDialog } from 'app/components/TransactionDialog';
 
 interface IRemoveLiquidityDialogV1Props {
   pool: AmmLiquidityPool;
@@ -248,7 +248,7 @@ export const RemoveLiquidityDialogV1: React.FC<IRemoveLiquidityDialogV1Props> = 
           )}
         </div>
       </Dialog>
-      <TxDialog
+      <TransactionDialog
         tx={tx}
         onUserConfirmed={() => props.onCloseModal()}
         onSuccess={props.onSuccess}
